Game start event and other changes
Also nuked the AI-compatible version and made the original AI-compatible from the start. I deem having to press J or otherwise manually plot a jump to be the lesser evil compared to the AI doesn't know what to do when suddenly given control of a previously player-controlled country with warp drives.